Project1/code/requirements.txt
Adafruit_Python_DHT==1.4.0
adafruit-circuitpython-neopixel==3.4.0
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
gpiozero==1.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pandas==0.25.2
picamera==1.8
python-crontab==2.3.9
python-dateutil==2.8.0
python-dotenv==0.10.3
plotly==4.2.1
retrying==1.3.3
RPi.GPIO==0.7.0
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
rpi-ws281x==4.1.0
